% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{Title}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{}

\item{K}{}

\item{M}{}

\item{numIter}{}
}
\value{
Explain return- return all values of class assignments as a vector
}
\description{
Title
}
\examples{
# Give example
set.seed(123)
X <- matrix(rnorm(100), nrow = 20)  # 20 points, 5 dimensions
# K-means with K = 3 clusters and random initialization
result <- MyKmeans(X, K = 3)
print(result)
 [,1]
 [1,]    1
 [2,]    1
 [3,]    3
 [4,]    1
 [5,]    1
 [6,]    3
 [7,]    3
 [8,]    1
 [9,]    1
[10,]    3
[11,]    3
[12,]    1
[13,]    3
[14,]    1
[15,]    3
[16,]    2
[17,]    3
[18,]    1
[19,]    3
[20,]    1
}
