% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{K-Means Clustering Wrapper Function}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{n by p matrix containing n data points to cluster.}

\item{K}{An integer specifying number of clusters.}

\item{M}{K by p matrix of cluster centers.}

\item{numIter}{number of maximal iterations for the algorithm, the default value is 100.}
}
\value{
Explain return- return all values of class assignments as a vector
}
\description{
K-Means Clustering Wrapper Function
}
\examples{
# Give example
set.seed(123)
X <- matrix(rnorm(100), nrow = 20)  # 20 points, 5 dimensions
# K-means with K = 3 clusters and random initialization
result <- MyKmeans(X, K = 3)
print(result)
 [,1]
 [1,]    1
 [2,]    1
 [3,]    3
 [4,]    1
 [5,]    1
 [6,]    3
 [7,]    3
 [8,]    1
 [9,]    1
[10,]    3
[11,]    3
[12,]    1
[13,]    3
[14,]    1
[15,]    3
[16,]    2
[17,]    3
[18,]    1
[19,]    3
[20,]    1
}
